feat(auth): setup jwt page redirection
TODO:
- refresh tokens

+++ apollo/resolvers.ts
@@ -1,3 +1,4 @@
+import Cookies from "cookies";
 import { sign } from "jsonwebtoken";
 import dbConnect from "../database";
 import Post, { IPostInput } from "../database/post.model";
@@ -12,15 +13,21 @@ const UserResolvers = {
   },
 
   Mutation: {
-    login: async ( _: any, { input: { _id, password } }: { input: IUserInput }) => {
+    login: async (
+      _: any,
+      { input: { _id, password } }: { input: IUserInput },
+      { cookies }: { cookies: Cookies }
+    ) => {
       await dbConnect();
-      // FIXME
       const generateTokens = () => {
         const key = process.env.SECRET_KEY ?? "ritrovo";
         const accessToken = sign({ _id }, key);
-        // const refreshToken = sign({ _id }, key);
-        // res.cookie("access-token", accessToken);
-        // res.cookie("refresh-token", refreshToken);
+        cookies.set("access-token", accessToken, {
+          httpOnly: true,
+          secure: process.env.NODE_ENV !== "development",
+          sameSite: "strict",
+          path: "/",
+        });
         return accessToken;
       };
 
@@ -40,24 +47,48 @@ const UserResolvers = {
 
 const PostResolvers = {
   Query: {
-    posts: async () => {
+    posts: async (_: any, __: any, { user }: { user: string }) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       return Post.find();
     },
 
-    postById: async (_: any, { id }: { id: string }) => {
+    postById: async (
+      _: any,
+      { id }: { id: string },
+      { user }: { user: string }
+    ) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       return Post.findById(id);
     },
 
-    postsByAuthor: async (_: any, { author }: { author: string }) => {
+    postsByAuthor: async (
+      _: any,
+      { author }: { author: string },
+      { user }: { user: string }
+    ) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       return Post.find({ author });
     },
   },
 
   Mutation: {
-    post: async (_: any, { input }: { input: IPostInput }) => {
+    post: async (
+      _: any,
+      { input }: { input: IPostInput },
+      { user }: { user: string }
+    ) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       const newPost = new Post(input);
       return newPost.save();
@@ -65,13 +96,24 @@ const PostResolvers = {
 
     updatePost: async (
       _: any,
-      { input: { _id, title, body } }: { input: IPostInput }
+      { input: { _id, title, body } }: { input: IPostInput },
+      { user }: { user: string }
     ) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       return Post.findByIdAndUpdate(_id!, { title, body }, { new: true });
     },
 
-    deletePost: async (_: any, { _id }: { _id: string }) => {
+    deletePost: async (
+      _: any,
+      { _id }: { _id: string },
+      { user }: { user: string }
+    ) => {
+      if (!user) {
+        return null;
+      }
       await dbConnect();
       return Post.findByIdAndRemove(_id);
     },
